Rename `epoch-recovery` tool to `epochs` and move methods
