--- cmd/util/cmd/epoch-recovery/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/onflow/flow-go/cmd/util/cmd/epoch-recovery/cmd"
-
-func main() {
-	cmd.Execute()
-}
