+++ cmd/util/cmd/epochs/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/onflow/flow-go/cmd/util/cmd/epochs/cmd"
+
+func main() {
+	cmd.Execute()
+}
