+++ cmd/util/cmd/root.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/viper"
 
 	checkpoint_list_tries "github.com/onflow/flow-go/cmd/util/cmd/checkpoint-list-tries"
-	epoch_recovery "github.com/onflow/flow-go/cmd/util/cmd/epoch-recovery/cmd"
+	epochs "github.com/onflow/flow-go/cmd/util/cmd/epochs/cmd"
 	export "github.com/onflow/flow-go/cmd/util/cmd/exec-data-json-export"
 	extract "github.com/onflow/flow-go/cmd/util/cmd/execution-state-extract"
 	ledger_json_exporter "github.com/onflow/flow-go/cmd/util/cmd/export-json-execution-state"
@@ -57,7 +57,7 @@ func addCommands() {
 	rootCmd.AddCommand(read_badger.RootCmd)
 	rootCmd.AddCommand(read_protocol_state.RootCmd)
 	rootCmd.AddCommand(ledger_json_exporter.Cmd)
-	rootCmd.AddCommand(epoch_recovery.RootCmd)
+	rootCmd.AddCommand(epochs.RootCmd)
 }
 
 func initConfig() {
